pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "glowtoolbox"
version = "0.1.0"
description = "GlowToolBox project"
authors = [{ name = "Your Name" }]
requires-python = ">=3.8"
dependencies = [
    # 图像处理
    "Pillow>=11.1.0",
    "pillow-avif-plugin>=1.4.6",
    "pillow-jxl-plugin>=1.3.2",
    "opencv-python>=4.11.0.86",
    "imagehash>=4.3.2",
    "psd-tools",
    "pyvips",
    "Wand",

    # 文档处理
    "PyMuPDF>=1.25.3",
    "pdf2image>=1.17.0",
    "openpyxl",
    "python-docx",

    # 压缩和文件处理
    "py7zr>=0.22.0",
    "patool>=3.1.3",
    "rarfile",
    "Send2Trash>=1.8.3",
    "watchdog",
    "fsspec",

    # 数据处理
    "numpy>=2.2.3",
    "pandas",
    "networkx",
    "msgpack>=1.1.0",
    "orjson",
    "ujson",

    # UI和交互
    "PyQt5",
    "textual>=2.0.4",
    "rich>=13.9.4",
    "tqdm>=4.67.1",
    "flet",
    "ttkbootstrap",
    "keyboard",
    "pyperclip>=1.9.0",
    "prompt_toolkit",
    "questionary>=2.1.0",

    # 系统和工具
    "psutil>=7.0.0",
    "pywin32",
    "portalocker",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.2",

    # 文本处理
    "chardet",
    "charset-normalizer",
    "diff-match-patch",
    "jaconv",
    "opencc-python-reimplemented",
    "pangu",
    "rapidfuzz>=3.12.1",
    "regex",

    # 日志和调试
    "loguru",
    "nanoid",

    # HTTP和网络
    "requests>=2.32.3",
]

[project.optional-dependencies]
ai = [
    "torch>=2.0.0",
    "onnx",
    "onnx2torch",
]

dev = [
    "pytest>=8.3.4",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "pip-tools>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["utils*", "nodes*", "scripts*"]
namespaces = false

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml", "*.env"]

[project.urls]
Homepage = "https://github.com/yourusername/glowtoolbox"
Repository = "https://github.com/yourusername/glowtoolbox.git"

[tool.black]
line-length = 100
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q" 